feedcard, answerfeedcard reactionbutton으로 교체
