+++ src/components/common/ReactionButton/ThumbsButton.tsx
@@ -3,17 +3,25 @@ import { postReactionOnQuestion } from 'api/api';
 import { ReactComponent as ThumbsUp } from 'assets/icon/thumbs-up.svg';
 import * as Style from './StyleThumbsButton';
 
-export type ThumbsButtonTypes = {
-  type : '좋아요' | '싫어요';
-  number?: number;
-  questionId?: number;
+export interface ThumbsButtonTypes {
+  reactionCount : number;
+  questionId: number;
+  type: '좋아요' | '싫어요';
   onClick?: () => void;
-  active?: number;
-};
+}
 
-const ThumbsButton = ({ number, questionId, type }: ThumbsButtonTypes) => {
-  const [likeNumber, setLikeNumber] = useState(number);
+const ThumbsButton = ({ reactionCount, questionId, type, onClick}: ThumbsButtonTypes) => {
+  const [likeNumber, setLikeNumber] = useState(reactionCount);
   const [active, setActive] = useState<'gray' | 'blue' | 'red'>('gray');
+  const [counting, setCounting] = useState({ up: null, down: null });
+
+  const handleCount = (type: 'up' | 'down') => {
+    if (!counting[type]) {
+      setCounting((prevCounting) => ({ ...prevCounting, [type]: 1 }));
+    } else {
+      setCounting((prevCounting) => ({ ...prevCounting, [type]: null }));
+    }
+  };
 
   const handleActiveClick = async () => {
     setActive('blue');
@@ -35,7 +43,9 @@ const ThumbsButton = ({ number, questionId, type }: ThumbsButtonTypes) => {
   return (
     <Style.Container active={active} onClick={handleActiveClick}>
       <ThumbsUp fill={active === 'blue' ? 'var(--blue)' : 'var(--gray40)'} />
-      <span>{type} {likeNumber}</span>
+      <span>
+        {type} {likeNumber}
+      </span>
     </Style.Container>
   );
 };
